+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## 1.0.1 - 2025-07-07 (Pathfinder 7.2.2)
+- Expose `getItem` function to CONFIG.itemSelectDialog.
+- Move types to submodule.
+
 ## 1.0.0 - 2025-07-05 (Pathfinder 7.2.2)
 - Initial development